+++ src/GuardClauses/GuardValidationDateTimeExtensions.cs
@@ -171,5 +171,48 @@ namespace PowerUtils.Validations.GuardClauses
                 throw new PropertyException(parameterName, ErrorCodes.MIN_DATETIME_UTCNOW);
             }
         }
+
+        /// <summary>
+        /// Throws an <see cref="PropertyException" /> if <paramref name="value"/> is equals to other value. Error code 'INVALID'
+        /// </summary>
+        /// <param name="_"></param>
+        /// <param name="value">Value to validate</param>
+        /// <param name="otherValue">Reference value for comparison</param>
+        /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
+        /// <exception cref="PropertyException">Exception thrown when value is equals to the other value</exception>
+        public static void IfEquals(
+            this IGuardClause _,
+            DateTime? value,
+            DateTime? otherValue,
+            [CallerArgumentExpression("value")] string parameterName = null
+        )
+        {
+            if(value == otherValue)
+            {
+                throw new PropertyException(parameterName, ErrorCodes.INVALID);
+            }
+        }
+
+
+        /// <summary>
+        /// Throws an <see cref="PropertyException" /> if <paramref name="value"/> is different to other value. Error code 'INVALID'
+        /// </summary>
+        /// <param name="_"></param>
+        /// <param name="value">Value to validate</param>
+        /// <param name="otherValue">Reference value for comparison</param>
+        /// <param name="parameterName">If not defined, the name of the variable passed by the <paramref name="value"/> parameter will be used</param>
+        /// <exception cref="PropertyException">Exception thrown when value is different to the other value</exception>
+        public static void IfDifferent(
+            this IGuardClause _,
+            DateTime? value,
+            DateTime? otherValue,
+            [CallerArgumentExpression("value")] string parameterName = null
+        )
+        {
+            if(value != otherValue)
+            {
+                throw new PropertyException(parameterName, ErrorCodes.INVALID);
+            }
+        }
     }
 }